+++ homematicip/device.py
@@ -1189,6 +1189,67 @@ class FullFlushBlind(FullFlushShutter):
 class BrandBlind(FullFlushBlind):
     """ HMIP-BBL (Blind Actuator for brand switches) """
 
+class BlindModule(Device):
+    """ HMIP-HDM1 (Hunter Douglas & erfal window blinds) """
+    def __init__(self, connection):
+        super().__init__(connection)
+        self.automationDriveSpeed = DriveSpeed.CREEP_SPEED
+        self.manualDriveSpeed = DriveSpeed.CREEP_SPEED
+        self.favoritePrimaryShadingPosition    = 0.0
+        self.favoriteSecondaryShadingPosition  = 0.0
+        self.primaryShadingLevel               = 0.0
+        self.secondaryShadingLevel             = 0.0
+        self.previousPrimaryShadingLevel               = 0.0
+        self.previousSecondaryShadingLevel             = 0.0
+        self.identifyOemSupported             = False
+        self.productId             = 0
+        self.primaryCloseAdjustable = False
+        self.primaryOpenAdjustable = False
+        self.primaryShadingStateType = ShadingStateType.NOT_EXISTENT
+        self.primaryCloseAdjustable = False
+        self.primaryOpenAdjustable = False
+        self.primaryShadingStateType = ShadingStateType.NOT_EXISTENT
+        self.profileMode = ProfileMode.MANUAL
+        self.userDesiredProfileMode = ProfileMode.MANUAL
+        self.processing             = False
+        self.shadingDriveVersion = None
+        self.shadingPackagePosition = ShadingPackagePosition.NOT_USED
+        self.shadingPositionAdjustmentActive = None
+        self.shadingPositionAdjustmentClientId = None
+
+    def from_json(self, js):
+        super().from_json(js)
+        c = get_functional_channel("SHADING_CHANNEL", js)
+        if c:
+            self.set_attr_from_dict("automationDriveSpeed", c, DriveSpeed)
+            self.set_attr_from_dict("manualDriveSpeed", c, DriveSpeed)
+
+            self.set_attr_from_dict("favoritePrimaryShadingPosition", c)
+            self.set_attr_from_dict("favoriteSecondaryShadingPosition", c)
+            
+            self.set_attr_from_dict("primaryCloseAdjustable", c)
+            self.set_attr_from_dict("primaryOpenAdjustable", c)
+            self.set_attr_from_dict("primaryShadingStateType", c, ShadingStateType)            
+            self.set_attr_from_dict("secondaryCloseAdjustable", c)
+            self.set_attr_from_dict("secondaryOpenAdjustable", c)
+            self.set_attr_from_dict("secondaryShadingStateType", c, ShadingStateType)
+
+            self.set_attr_from_dict("primaryShadingLevel", c)
+            self.set_attr_from_dict("secondaryShadingLevel", c)
+
+            self.set_attr_from_dict("previousPrimaryShadingLevel", c)
+            self.set_attr_from_dict("previousSecondaryShadingLevel", c)
+
+            self.set_attr_from_dict("identifyOemSupported", c)
+            self.set_attr_from_dict("productId", c)
+
+            self.set_attr_from_dict("profileMode", c, ProfileMode)
+            self.set_attr_from_dict("userDesiredProfileMode", c, ProfileMode)
+
+            self.set_attr_from_dict("shadingDriveVersion", c)
+            self.set_attr_from_dict("shadingPackagePosition", c, ShadingPackagePosition)
+            self.set_attr_from_dict("shadingPositionAdjustmentActive", c)
+            self.set_attr_from_dict("shadingPositionAdjustmentClientId", c)
 
 class LightSensor(Device):
     """ HMIP-SLO (Light Sensor outdoor) """